+++ lib/dynamoid/persistence/update_fields.rb
@@ -20,7 +20,7 @@ module Dynamoid
         UpdateValidations.validate_attributes_exist(@model_class, @attributes)
 
         if @model_class.timestamps_enabled?
-          @attributes[:updated_at] ||= DateTime.now.in_time_zone(Time.zone)
+          @attributes[:updated] ||= DateTime.now.in_time_zone(Time.zone)
         end
 
         raw_attributes = update_item